Move fol and logicalform to logical_representation
